Fix for nested dirs in bundler-friendly URLs
During review of #14135 some of the changes made TARGET_JS_NAME to be a relative path instead of a basename.

This way, when someone specified `-pthread -o out.js`, it worked fine, but if someone used `-pthread -o nested/dir/out.js` then the `nested/dir/out.worker.js` would have a reference to `./nested/dir/out.js` not to `./out.js` as it should for a file alongside it.