Add a simple one-entry-per-line file format for specifying symbol lists
For options such as `EXPORTED_FUNCTIONS` that take lists we now support
a simple file format that is just one line per symbol.   It doesn't
require any escaping or punctuation.

This is much easier to use maintain and is in line with other compiler
toolchain options used by gcc and ld.  For example ld's
`--retain-symbols-file=filename` or wasm-ld's
`--undefined-file=filename`.

This should be especially useful for options like `ASYNCIFY_ADD` or
`ASYNCIFY_REMOVE` which often involvea lot of quoting (because they
include things like spaces and commas).