Auto-export setThrew for emscripten exceptions (#14153)
Normally if exceptions are enabled, we have exceptions thrown and caught
etc., and they will depend on setThrew. And then we will link it in. However,
if the code only contains invokes, but no exception catching, then there is
no library support that tells us to include it. To fix that, always include the
function when emscripten exceptions are enabled.

I don't think this can regress code size, as setThrew will almost certainly
be used anyhow, unless there are also no invokes at all. And in that case,
metadce can remove it.

Fixes #14078