Fix .bat launcher to allow special characters
Using $* inside the if/else blocks here was confusing the bat file
processor making it impossible to support commands that contains braces
and exclamation points.

There was an attempted fix made in #13723 which seemed hopefull but it
prevented the exclamation point.  Hopefully this will fix both cases.

Thanks to Mysoft on discord for the suggestion.

Fixes: #14063