Rather than disabling the bullet test, temporarily remove -g from the build. (#14264)
We are seeing more failures related to crash that is currently
occuring when bullet is built with `-g`.  This changes the way
we temporarily work around this issue.

Also it seems that mpg123 is now also experiencing the same
crash.  All the more reason to get this fixed upstream ASAP.

See: #14255