Allow unsetting default environment vars
User code can set environment variables in ENV, which are then merged
with Emscripten's defaults. However, there is currently no way to unset
a variable entirely: if it's not set in ENV, it will just keep the
default value. For instance, you cannot unset LANG.

This change makes it possible to explicitly unset variables by setting
the corresponding entry in ENV to undefined.

Note that the existing behavior for undefined in ENV is to represent it
as the string "undefined" in C. That's not a useful behavior (anyone who
actually wants that effect will pass the JavaScript string "undefined"
instead), so we are free to repurpose it.