Do not emit intermediate debug info by mistake in the final output (#14154)
In some build modes (certain asyncify flags) we forgot to remove debug info
at the very end, if we only needed it in the middle.

Note that this only affects the names section - function names. We did not
have any issues with DWARF, which has explicit stripping code already, and
that is heavily tested (as DWARF can be huge) as correct.

The fix here is by tracking the reasons why we need interim debug info, and
noting when we no longer need it. We can then strip it out at the end, if so.

We also stop emitting it as soon as possible, which makes the build slightly
more efficient. Also, this can avoid running wasm-opt just to strip the debug
info if it's already been removed by a previous command, which is potentially
a large speedup in a big project in an unoptimized build (as in an optimized
one there will be another command that strips it anyhow). To achieve that,
this adds tracking of whether the last command stripped debug info.

Fixes #14143