Fix browser.test_async_2 (#14286)
The test regressed after an LLVM change. The LLVM change removes the
recursion (!), which then lets binaryen inline it, which changes the stack
traces radically. To work around that, do not run the binaryen optimizer.

I looked at the history of the test to make sure the optimizer wasn't a part
of what is tested, and it seems not. The test originates in the emterpreter,
so it was not even an asyncify test originally. But even there, the test checked
for trampling of locals due to the async transform, and that would be the
same in an opt build as otherwise (that is, the same local storing/loading code
is used in all cases; if the optimizer breaks that, it would be a general optimizer
bug and not specific to asyncify).

Also mark the function as noinline, which is something the test depends on
heavily.