cmake: Default to installing into sysroot
Assuming the user/project does not specify a CMAKE_INSTALL_PREFIX
to install to we can and should default to installing into the
emscripten sysroot.

Also, add a test that uses `cmake --install` to install a library and
the `finc_package` to find the library config.

This confirms that #6595 is fixed even though technicall the fix landed
in in #14291