Reland: Warn if link-only settings are used in compile-only mode (#14182)
Originally landed in #14042 and then reverted in #13884.

Since then I landed a change that makes false positives impossible by
asserting if any link-only setting is accessed before the link phase.
This means that there are cannot be any setting that is used during the
compile phase that is *not* part of COMPILE_TIME_SETTINGS.  If such a
setting existed we would hit an assert during testing.  (Obviously this
depends on our test coverage being good).