Stop doing extra optimization work in standalone mode (#14338)
Before this PR we would try to remove unneeded imports, as they might cause
a wasm not to run in a WASI VM if it were nonstandard. However, I think that
was only an issue when bringing up standalone mode (as we converted nonstandard
things to WASI). Today, I verified that building hello world (even with C++ iostreams)
in -O0 will run in both wasmtime and wasmer without issue.

While there may be cases in which there are nonstandard imports, it is best
to let the user explicitly run the optimizer to work around that, as otherwise it
adds complexity (in particular, as sbc100 noticed, if we automatically run the
optimizer even in unoptimized builds, it means that
ERROR_ON_WASM_CHANGES_AFTER_LINK will always error in
standalone mode).