Use `-lxxx` internally rather than full library path
Since we already supply the correct `-L` path we can rely on wasm-ld to
find, for example, libc.a by simply pasing `-lc` rather than its full
path.

This is in line with what gcc and clang do.

This is almost NFC, except in that case where the user supplies and
library path via `-L` which happens to contains a file called `libc.a`.
In this case, as with clang or gcc, that file will be chosen over the
system one.

This makes the linker command lines shorter, more readable, more inline
with other compilers.