Avoid using get_building_env in test code
Rather than using the internal `building.get_building_env` function
we can just make sure we run make via `emmake` (or that we use
`emcmake` or `emconfigure`).

If we do use `emcmake` or `emconfigure` then any environment varaibles
should be internalized during that phase so we don't need to pass a
custom envrionment to the `make` call.

This also allows us to remove the optional param which only existed for
the sake this one caller in test code.