Fix ASYNCIFY + EXIT_RUNTIME check (#14418)
This check had a typo and was doing `Asyncify.state === Asyncify.State.None`
rather than ``Asyncify.state === Asyncify.State.Normal`.

This meant it was always firing if any ASYNCIFY program happened to
actually exit the runtime.  Note that ASYNCIFY programs almost never
exit do this should actually effect most programs.  This is because any
actual use of async functions during main will disable exiting of the
runtime (See #14417).

However, even after fixing the typo this assert still fires in all cases
because `Normal` is the state adter the rewinding is done.

Instead I use a an explict debug state to indicate that asyncify is
currently disabled.

I added a null version of the test that doesn't use ASYNCIFY during exit to be
sure this fires only when it should.