Fix asan.test_utf8_textdecoder (#14428)
The test has some loops that try to remove certain utf8 characters
from the end (whitespace? I don't know enough utf8). The loops were
missing a check on the string size not being 0.

This bug existed before #14399, but before that PR we always used
strings of size 8, and apparently were lucky enough to not run into
a run of 8 characters that we want to remove. With that PR, the bug
unsurfaced as we try various lengths, even 1.

The asan test suite will be green again after this.