Switch Firefox testing to headless mode, and disable GL testing there (#14434)
I cannot reproduce the errors in #14401 locally, so I investigated on
the CI side. Looking at many failures, they are all in tests that use GL. And so a
theory is that the odd graphics setup on CI is to blame somehow - perhaps the
GL driver crashes firefox randomly. An update to those GL drivers could
explain why this happens so frequently now.

This PR switches us away from using that odd graphics setup, and instead to
use headless mode. Headless mode does not support GL - that is why we
avoided it before - but it seems like now we have no choice but to avoid GL
tests on firefox. That must wait until firefox fixes https://bugzil.la/1375585

To confirm the theory here, I ran 8 runs of CI, and all tests passed, so I think
this makes us stable again. The cost of not doing GL testing is significant,
but I don't see an alternative.

This PR marks more tests as needing GL hardware, that were missing that
tag.

This also disables emrun and posix tests. (That's just a very small number of
tests, so it doesn't seem so bad.) The issue there is that something about
headless mode doesn't mix well with wanting to start another firefox
process, see the comment in the PR. It might be worth experimenting with
killing the firefox process or something like that, but I think this could land
first.

This also returns us to testing Firefox dev edition, as before.