Remove unused dependency from `emscripten_start_fetch__deps`
The function is a little unusual in that is provided by native code in
USE_PTHREADS builds but by JS library code in non-threaded builds.

In neither case does the JS fetch code depend this symbol.  A call
to this function exists in `system/lib/fetchemscripten_fetch.cpp` but
that should be resolved using normal native->native or native->JS
linking which does not require any special handling.

See: #14313