Limit LIBRARY_DEBUG to functions that can be exported to C/C++ (#14527)
I think the indent with this macro was that it was designed to trace at
the native/JS boundary.  As it stands it traces all JS library functions
including those that take arbitrary JS objects as arguments.  This cases
`prettyPrint` to crash when trying to print those arguments.  For
example enabling LIBRARY_DEBUG with MAIN_MODULE will currently cause the
call to `$relocateExports` to crash trying to print the `exports` object
(which is an object with a null prototype and therefor without a
`toSrring()` method).