Fix use of shared state in test multiprocessing
In #14536 I moved away from solely using environment variables to
control testing options and towards using commmand line args.  These
command line args set global state in the same way that the env vars do.
However, due to the way multiprocessing works on windows this global
state is not shared with child processes.

On windows the 'spawn' method is used to create children whereas on
unix the 'fork' method us used.

With this change alwasy inject the environment varisable even when not
running `main`.  I also set those environment variable based on the
command line options so they will always be shared with the child
proccess.

In the fugure we should probably transmit out state/option explcitly to
out children but for now this fixes the issue we have been seeing on the
windows roller.