Fix use of shared state in test multiprocessing (#14556)
In #14536 I moved away from solely using environment variables to
control testing options and towards using command line args.  These
command line args set global state in the same way that the env vars do.
However, due to the way multiprocessing works on windows this global
state is not shared with child processes.

On windows the 'spawn' method is used to create children whereas on
unix the 'fork' method us used.

With this change always process the environment variables even when not
running `main`.  I also set those environment variable based on the
command line options so they will always be shared with the child
process.

In the future we should probably transmit out state/option explicitly to
out children but for now this fixes the issue we have been seeing on the
windows roller.