Make alignMemory into a library function
This allows it to be used in MINIMAL_RUNTIME which can end up depending
on it in asan builds due to asan using mmap (which calls mmapAlloc).

This fixes the current failure on the waterfall of:
  asan.test_dyncall_specific_minimal_runtime

Also remove the default for the second argument here.  There are only
two call sites and both of them specify an alignment.