Remove use of addAtMain is webidl_binder.py
Support for `addAtMain` in application without a main function was
removed in #13901 but I missed that fact that webidl_binder.py was doing
this.

This change replaces `addAtMain` with `addOnInit` and also makes the use
of `addAtMain` in applications withoiut a `main` into an error so that
the problem (and solution) should be clear.

Update webidl binder test to only use sync compilation when needed, so
that code in question actually gets tested.

This change also reduces code size by not including `addAtMain` at all
in applcications without a `main`.

Fixes: #14583.