Adjust initial memory to handle asan shadow size
Without this change any user of ALLOW_MEMORY_GROWTH
with asan would get a rather cryptic error message from
the linker:

 wasm-ld: error: initial memory too small, 278278624 bytes needed

This is because asan increases the memory requirements in
proportion to the max memory, and the max memory defaults
to 2Gb with `ALLOW_MEMORY_GROWTH` enabled.

Also, remove check for `MAXIMUM_MEMORY` == -1 which is no longer
a thing since #14088.

Fixes: #14621