Optimize memcmp (#14702)
This does similar things to what we already do for memcpy etc.,
doing word-sized operations where possible.

On artificial benchmarks this is up to 3.7x faster. This is motivated
by a real-world benchmark where memcmp is very high in the profile.

For consistency, also apply EMSCRIPTEN_OPTIMIZE_FOR_OZ as we
do for memcpy etc., to not unroll loops in clang in that case.