Explicitly create the proxied main thread as detached
Allowing musl to know whether it can immediately dispose of the
thread resources once the thread is exited, instead of keeping
it around (for a possible `pthread_join` call, which will never
occur for this thread).

This change can probably be considered non-functional, because
the proxied main thread was previously disposed anyhow during
`PThread.terminateAllThreads`.

Resolves: #9467.