Warn when we map unrecognized library files to `-l` flags.
This is rather an odd feature and I think we should notify users if they
are relying on it.  Honestly I doubt many are.

See #14689 for an example of how confusing it can be when it goes wrong.

Also, make sure we strip off the full suffix in the case of `.so.1.2.3`
suffixes.