Fix for wasmOffsetConverter with MODULARIZE
I tried to fix this in #13236 but I don't think it was ever working
correctly.  In MODULARIZE mode `wasmOffsetConverter` was not being
defined within the module itself (where it is used).  his change moves
the contruction of `wasmOffsetConverter` and `wasmSourceMap` so they
always occur inside the module.

I've also improved the testing of `wasmOffsetConverter` so we actually
verify that it works in `tests/core/test_return_address.c`.