Run node with `--unhandled-rejections=throw`. NFC
This mode means that unhandled promise rejections will be thown as
exceptions which means that the process will exit with a non-zero return
code.  This is the default mode in v15 and above but we still use v14
and we want this behaviour when running our tests to make sure we don't
silently ignore any unhandled rejections.

See https://nodejs.org/api/cli.html#cli_unhandled_rejections_mode.