file_packager.py: Only store the audio file property when true (#14798)
In the .js generated by file_packager.py which currently produces the following attributes for each file,

{"filename":"<path>","start":<start>,"end":<end>,"audio":0}

(exemple here) only store the audio attribute when it's set to 1. Otherwise assume that it is 0 (which should be the case for the vast majority of files). Resulting in,

{"filename":"<path>","start":<start>,"end":<end>}

This saves 9 bytes per file, or in the above linked example about 10kB for 1100 files.

Related to #14695