Fix pthreads+dylink+long .so names (#14854)
In pthreads mode, TextDecoder is wrapped in a special way. The logic there
receives buffers and copies them. It asserts the buffer is a Uint8array. The
dylink code hits that assertion as it creates a Int8Array and sends it there.
We could make the wrapper support signed views too, but it seems simpler
to just use the same unsigned view there, which is clearer anyhow.

Fixes #14833