Support Sanitizers + Asyncify (#14864)
The problem here was that the asan_c_load_4 etc. methods, that are tiny
functions in C that we let ASan instrument, and then JS calls into, were getting
processed by Asyncify. And we end up calling those asan_c_* methods
during critical Asyncify times like as we begin to rewind - times when it is
not valid to call into instrumented code, or we assert. To fix that, simply
ignore those specific methods in Asyncify - we don't need to process them
anyhow, as they cannot pause.

One test remains broken, test_asyncify_lists_onlylist_d, which I left as
a TODO. That test does something invalid, and it is expected to fail, but
it fails in a different way somehow. The details get complicated due to how
ASan modifies things. Let's leave #14807 open for that.