Allow test to override name of output JS file. NFC
This is useful in the posixtest suite where the source
filenames are things like `pthread_key_create/1-2.c` resulting
in a default output name of `1-2.js` which is not very useful
when looking at `ps` output of running tests.

With this change the JS files can match the test name
(e.g. test_pthread_testcancel_1_1.js).