Mention that adding oneself to AUTHORS is optional (#14893)
I have received guidance that AUTHORS is not strictly needed. Like most
projects, we can just rely on git history if we ever need the list of
contributors.

However, since we have the file, we should keep it around for historical
purposes. Just mark it as optional.

We should also disable the bot that auto-suggests people add themselves.

fixes #14354