Remove stale api_items.py. NFC
This file can (and should) just be generated whenever its needed
using get_api_items.py.

Having it checked into git means that we `git grep` for anything
in the API will often show result from this file and we don't want
to have to worry about keeping the checked in copy up-to-date.