Run core/test_safe_stack.c under pthreads + node. NFC
We already run this test with pthreads in test_browser.py but its useful
to run in this mode under node as well where the threading model is not
identical.

Also, add `assert_all=True` when running these tests as we want to march
all of the patterns, not just one.