[WIP] Add thread exit hooks
The first current issue is that in-thread.profile should not be necessary, but
for some reason it is. thread.profile is written in the thread exit hook so it
should include functions like `thread` and `foo`, but it does not while
in-thread.profile does. This is mysterious because thread.profile is written
later than in-thread.profile and on the same thread, so one would expect it to
contain a superset of the functions in-thread.profile contains. Is it possible
that the thread migrated between workers?

The second current issue is that the atExit handler runs too early so there a
bunch of shutdown functions including fflush and its dependencies that it does
not include. This is not a serious issue because it's not too onerous to list
these functions manually on the wasm-split command line, but that is a fragile
solution.