Fix test_dylink_spaghetti
This test was inadvertently relying on an bug in llvm which was recently
fixed: https://github.com/emscripten-core/emscripten/issues/13773

In this test the class `Class` is defined in both the main and the side
module.  If you run this code as is on the desktop the version of
`Class` defined in `main.cpp` will always win the `side init` will never
get printed.

The fix is to give the class in the side module a different name (and
alternative would be to put it in an anonymous namespace).

This change should allow the llvm fix to roll in:
https://reviews.llvm.org/D108413