Remove EmccOptions.tracing and thread_profiler. (#14956)
These options have the same effect as the equivalent settings.  Having
them exist as both options and settings can lead to bugs where the
option object is checked but not the setting.

For example this change fixes a bug where `-D__EMSCRIPTEN_TRACING__` was
passed the compiler when `--tracing` was set but not when
`-sEMSCRIPTEN_TRACING` is passed on the command lind.