Don't call emscripten_tls_init from static constructor. NFC
This change makes handling of `emscripten_tls_init` the
same on all threads, including the main thread and/or library
loading thread.

This is precursor to a change that makes use of the return
value of emscripten_tls_init to do TLS relocations so making
the call explicit is required.

I've also removed the use of EM_ASM from library_pthread.c.