Fixing incorrect argument list in JSimplementation
Arguments which were arrays were not getting an array pointer in the argument list for only the JSimplementation glue code. This was due to passing raw.type.name to type_to_c which did not allow for the square brackets to be processed. Note that since this fix pertains to the generated glue code, testing without the fix in place will throw an error in the tester due to mismatched argument types.