Fix asan.test_fs_trackingdelegate by fixing open() call (#15021)
It turns out that when O_CREAT is given to open(), one must
also provide the flags with which to create the file. Not providing
that varargs param is apparently undefined behavior, and the
musl open() method will end up reading a vararg that does not
exist, which leads to a null deref.