Avoid compile error on test_aligned_alloc
A recent commit taught clang that the first parameter to aligned_alloc should be
a power of two alignment, and since then clang emits a warning (which we upgrade
to an error) when that argument is a constant that is not a power of two. Since
the test actually wants to test the function behavior in that case, ignore the
error.