Fix checks for nodejs and/or the require function
The check for `require` was being applied even in the case that it
wasn't needed.

The check for node, when running under the shell environment was out of
date with the node check that we use to define ENVIRONMENT_IS_NODE and
should not have been checking for require (which is not always present
under node these days).

Fixes: #15022