Only install `unhandledRejection` handler if needed
We install this handler on node so that we can set the exit code
correctly.  However this is done by default on node 15 above and not
installing this handler makes reading error messages much clearer
because it preserves the stack traces.

Also add a test for the specific behaviour of this setting.