Fix tests and assertions after llvm change to handle SjLj+exceptions
As of https://reviews.llvm.org/D109670, exception handling support
in llvm will now generate calls to `emscripten_longjmp` when doing
exception handling (so that it can re-throw 'longjmp' exceptions).

Because we build our system libraries with SUPPORT_LONGJMP enabled
this means that the `std::__terminate` in libcxxabi now contain a
reference to `emscripten_longjmp` when built with exceptions
enabled.