Remove welcome message bot (#15084)
The bot used to be very important while we required people to add
themselves to AUTHORS - it was a useful reminder for that. But we've
removed that requirement. Now the bot just says a friendly message,
but we've gotten negative feedback on that,

#15071 (comment)

I think I agree that a content-less message like that kind of sends an
unclear signal, as the normal expectation is to get someone to review it.
That is, the bot message sounds defensive. Let's just remove it.