Default to EXIT_RUNTIME=1 when building with ASan (#15083)
We were already setting `EXIT_RUNTIME=1` under LSan because without this
leaks are not reported.  Since ASan includes LSan by default it makes
sense to do that same for ASan.

Because this change also effectivly enables leak detection in the `asan`
test suite it also includes a bunch of fixes for memory leaks.

The code in `emcc.py` that has been moved depend on `EXIT_RUNTIME` and
so must be run only after this setting has been finalized.

As part of this change I found a couple of issues related to
`EXIT_RUNTIME` when combined with other flags:

https://github.com/emscripten-core/emscripten/issues/15080
https://github.com/emscripten-core/emscripten/issues/15081

Regarding memory that leaked from JS and/or system libraries there in
some cases we want to ignore the leaks rather then adding code to free
the memory and in this case I've used `emscripten_builtin_malloc`.