Don't mark a thread a joinable when it calls exit()
The previous behavior was that ExitStatus was thrown which ends up
calling emscripten_thread_exit and marking the current thread as
join-able.

This this bug was masking several failures in the posixtest suite
which have always be failing but were not being recognized as such.

Note that we have to use the message queue rather than postMessage
to deliver the `exitProcess` message since postMessage will not interrupt
the main thread if its blocking on pthread_join.

Modify the existing test so that it fails if pthead_join even returns.

Fixes: #14993