Add `__noleakcheck` library attribute. NFC
This library function attribute signals that the function should be
wrapped in `withBuiltinMalloc` but is a no-op for non-sanitizer builds.

This is small win for code side when santizers are not enabled and
it enables enture functions to be easily wrapped without creating
extra lambdas/indentation in the code.

I was looking into wrapping __pthread_create_js and was wanting
something like this that would be minimally intrusive.