Move more dlfnc functionality from JS to native code. NFC
This is preparation for sharing state of loaded libraries
between threads.

One thing we not long implement is actually removing a loaded
library using `dlclose`.  Basically we assume the reference
count of a loaded library never gets to zero.  This seams reasonable
given the musl itself doesn't handle this case. See:
https://github.com/emscripten-core/emscripten/blob/0d5d7d38e0af59d19be8edad817b086a766e7869/system/lib/libc/musl/src/ldso/dlclose.c#L6-L9

All that happens in `dlclose` is a check for the validity of the
handle, no ref count is updated and no libraries are actually
unloaded.