Restore timeout on running js code during testing (#15160)
These days we can just use timeout parameter of subprocess.run so it
much simpler. No need for polling or any kind of watchdog.

This means that tests that never return will now show up as failures in
CI.  Currently the entire test run just never returns and its not
possible to see which tests (if any) were the responsible for the
timeout.