Fix inability to load WASM files when a slash is in the JS query/hash (#15152)
scriptDirectory is not always correctly set when the URL of the JS
file contains a querystring (preceded by a ?) or a fragment identifier
hash (preceded by a #) because these can contain a slash (/), which is
getting mistaken for part of the path.

Severing JS files with a querystring or hash can be useful for a variety of
reasons, including cache busting. (Even base64 encoded strings may
contain a slash.)

This PR removes any querystring or fragment hash from scriptDirectory.