Avoid re-using thread IDs. NFC
I'm not clear this fixes any known issues but it certainly makes
debugging pthread issues easier when each thread has a unique and
non-reusaable ID.

I believe its also safer then resuing the address of the pthread
structure here since that can (and often is) re-used for other
things or indeed other new threads (so it not unique for the
life of the program).

I also noticed that this field (and serveral others) are no longer
needed in `struct_info_internal.json`.