Update syscall naming such that native symbol and JS symbols match
Previously we were using obscure internal names such as __syscall221
and then mapping this to more meaningful names when importing them.
We were using the `import_name` to do this.

However, it turns out we can actually give these symbols meaningful
names internally and avoid this mapping completely.

As part of this I had to use the `__syscall_` prefix rather than
the `__sys_` prefix we had been using becuase `__sys_open` is already
used as a macro inside of musl.