Remove usage of syscall macros that we don't support
Followup to #15193.  With #15193 I removed the unused iplemenations,
this change always us to not even define these missing syscall numbers
at all.

This change removes the only place in the codebase where `SYS_xxx`
macros are assumed to be numbers.  This change is part of a bigger
change to switch to using meaningful string names for `SYS_xxx` macros
rather than numbers.