Fix for dlopen(NULL)
To match how musl works we ensure that the list of dso's always starts
with the main module (AKA `__main__`, AKA `dlopen(NULL)`.

Add more assertions to test_`dlfcn_self` to ensure we catch this in the
future.  It turns out the `dlopen` was returning NULL but that
`dlsym(NULL)` was then doing the right thing by coincidence.

Because the native version of `dlopen` already both check for null
filename and handle it there is no need to the JS version to handle
NULL.