Move trivial unistd tests to other (#15268)
We don't need to run them in all opt modes. They do depend on JS, so
the opt mode can matter in theory, but to really test that we need to run
closure - which we have on a few strategic tests already. This just moves
the trivial tests that don't use closure or any interesting decorators etc.

(In general I think we should move more things out of core into other
in order to avoid wasted testing and speed up our CI, but we'll need to
consider each test on a case by case basis.)

This also renames the wasmfs test to have a wasmfs_ prefix.