Remove EMIT_EMSCRIPTEN_METADATA settings
We have been warning about the deprecation of this setting
for about a year now (since 2.0.9).

As explained in #12231, the major use cases for this metadata
are no longer valid since the embedder should on longer need to
know the memory layout of the module (since the memory and table
are both created and setup within the wasm module these days).

Fixes: #12231