Update Embind+GC integration
Updates defunct integration from #8474 to latest version of the spec that renamed FinalizationGroup -> FinalizationRegistry, and changed the callback to take one value at a time instead of an iterator.

Additionally, even though this functionality is now implemented in all major browsers, it has zero guarantees about reliability, and the callback might fire very late or not fire at all and still stay spec-compliant.

As such, as suggested in #11436, the integration is used to not only free the found leaked instances, but also print a warning in ASSERTIONS mode that helps developer to find the original leaked instance and encourages to .delete() it manually.

Resolves #11436.