Merge `addFunction` and `addFunctionWasm`. NFC
In updating the assert text in `addFunction` I noticed that we don't
really need the internal `addFunctionWasm` anymore and we can just merge
them for simplicity.

`addFunctionWasm` was add as an internal function and never documents so
should be safe to remove.

The assertion from `addFunction` is more precisely handled the assert
in the exception handling block:

```
    assert(typeof sig !== 'undefined', 'Missing signature argument to addFunction: ' + func);
```

This handles that case where addFunction is given a JS function but no
signature.