Add deps to `tzset`
These deps are on native code so will not actually cause the symbols to
be included, that is the job of the deps_info.js entry for `tzset`.
However, it does mean that if the deps_info entries don't do their job
for whatever reason we get a nice link time failure rather than a
runtime failure.